+++ src/main/java/com/example/FreeSoulsApi/repositories/ArmorsRepository.java
@@ -3,5 +3,5 @@ package com.example.FreeSoulsApi.repositories;
 import com.example.FreeSoulsApi.entities.Armor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArmorsRepository extends JpaRepository<Armor, Integer> {
+public interface ArmorsRepository extends JpaRepository<Armor, Long> {
 }
