+++ src/main/java/com/example/FreeSoulsApi/repositories/WeaponsRepository.java
@@ -3,5 +3,5 @@ package com.example.FreeSoulsApi.repositories;
 import com.example.FreeSoulsApi.entities.Weapon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WeaponsRepository extends JpaRepository<Weapon, Integer> {
+public interface WeaponsRepository extends JpaRepository<Weapon, Long> {
 }
